+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/ui/TestTypesAdapter.java
@@ -53,8 +53,19 @@ class TestTypesAdapter extends ArrayAdapter<TestInfo> {
         View rowView;
 
         rowView = inflater.inflate(R.layout.row_type, parent, false);
-        ((TextView) rowView.findViewById(R.id.textUnit)).setText(
-                testInfo.getName(mActivity.getResources().getConfiguration().locale.getLanguage()));
+
+        if (testInfo.isGroup()) {
+            rowView.findViewById(R.id.textGroup).setVisibility(View.VISIBLE);
+            if (testInfo.getIsDiagnostic()) {
+                ((TextView) rowView.findViewById(R.id.textGroup)).setText("Experimental");
+            }
+            rowView.findViewById(R.id.typeLayout).setVisibility(View.GONE);
+        } else {
+            rowView.findViewById(R.id.textGroup).setVisibility(View.GONE);
+            ((TextView) rowView.findViewById(R.id.textName)).setText(
+                    testInfo.getName(mActivity.getResources().getConfiguration().locale.getLanguage()));
+        }
+
 
         return rowView;
     }